Update 07-find.md
I could not figure out why the command "cut -d : -f 2" was provided as an option. There is also no sense for where the command is called from. This extra example provides a location anchor and a hint about the semicolon syntax. Instead of a hint, the semicolon separator could be noted during the LittleWomen.txt example, where it is seen for the first time.